backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.9.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.6
